Add files via upload
This uploaded file is a revision of the previous version. The main update involves refining the ontology related to sodium channel activity and potassium channel activity, making the classifications more specific and aligned with their physiological functions.
- sodium channel activity (GO:0005272) changed to voltage-gated sodium channel activity (GO:0005248)
- potassium channel activity (GO:0005267) changed to voltage-gated potassium channel activity (GO:0005249)
- leak channel activity remain the same, because in GO, there is only potassium ion leak channel activity. The ions involved in leak channel are Na+ and K+.
These improvements aim to enhance the accuracy and clarity of the model's representation of ion channel behaviour.